+++ src/App.tsx
@@ -2,11 +2,14 @@ import { ThemeProvider } from 'styled-components'
 import { defaultTheme } from './styles/themes/default'
 import { GlobalStyle } from './styles/global'
 import Transactions from './pages/Transactions'
+import { TransactionsProvider } from './contexts/TransactionsContext'
 
 export function App() {
   return (
     <ThemeProvider theme={defaultTheme}>
-      <Transactions />
+      <TransactionsProvider>
+        <Transactions />
+      </TransactionsProvider>
 
       <GlobalStyle />
     </ThemeProvider>
